requirements.txt
argcomplete==3.4.0
asttokens==2.4.1
backoff==2.2.1
censys==2.2.12
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
dnspython==2.6.1
executing==2.0.1
fake-useragent==1.5.1
idna==3.7
ipaddress==1.0.23
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prettytable==3.10.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.0.3
requests==2.32.3
rich==13.7.1
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
